Rename MultiSigCommand to MultiSigTransaction
