+++ multisig/index.ts
@@ -1,5 +1,5 @@
-export * from './command';
 export * from './erc20';
 export * from './execution';
 export * from './ownership';
+export * from './transaction';
 export * from './transfer';
